% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_chatgpt}
\alias{shiny_chatgpt}
\title{Simple chatGPT app}
\usage{
shiny_chatgpt()
}
\value{
없음
}
\description{
chatGPT를 이용한 채팅 및 이미지 생성을 위한 Shiny 앱 호출
}
\examples{
\dontrun{
 library(shinyGPT)

 ## chatGPT를 이용한 채팅 및 이미지 생성기(Shiny Web Application) 호출
 shiny_chatgpt()
}
}
\seealso{
\code{\link{chat_completion}}, \code{\link{draw_img}}, \code{\link{draw_img_variation}}
}
\author{
유충현
}
